+++ src/main/java/nextstep/courses/domain/session/inner/FreeSessionType.java
@@ -1,10 +1,14 @@
-package nextstep.courses.domain.session;
+package nextstep.courses.domain.session.inner;
 
+import nextstep.courses.domain.session.Session;
 import nextstep.payments.domain.Payment;
 
-class FreeSessionType extends SessionType {
+public class FreeSessionType extends SessionType {
+
+    private static final String TYPE = "FREE";
 
     public FreeSessionType() {
+        this.sessionType = TYPE;
         this.maxEnrollment = -1;
         this.fee = 0;
     }
