+++ src/main/java/nextstep/courses/domain/session/inner/Image.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain.session;
+package nextstep.courses.domain.session.inner;
 
 public class Image {
     private String url;
