+++ src/main/java/nextstep/courses/domain/session/inner/SessionDate.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain.session;
+package nextstep.courses.domain.session.inner;
 
 import java.time.LocalDate;
