+++ src/main/java/nextstep/courses/domain/session/inner/SessionType.java
@@ -1,0 +1,13 @@
+package nextstep.courses.domain.session.inner;
+
+import nextstep.courses.domain.session.Session;
+import nextstep.payments.domain.Payment;
+
+public abstract class SessionType {
+
+    protected String sessionType;
+    protected int maxEnrollment;
+    protected long fee;
+
+    public abstract boolean enroll(Payment payment, Session session);
+}
