+++ src/test/java/nextstep/courses/domain/session/SessionTest.java
@@ -4,10 +4,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDate;
 
+import nextstep.courses.domain.session.inner.FreeSessionType;
+import nextstep.courses.domain.session.inner.PaidSessionType;
+import nextstep.courses.domain.session.inner.SessionStatus;
 import nextstep.payments.domain.Payment;
 import org.junit.jupiter.api.Test;
 
-class SessionTest {
+public class SessionTest {
     public static final Session SESSION1 = new Session(1L, 1L, "free", new FreeSessionType(), LocalDate.of(2023, 10, 1), LocalDate.of(2023, 10, 31), SessionStatus.OPEN, 0);
     public static final Session SESSION2 = new Session(2L, 1L, "paid", new PaidSessionType(10, 10000), LocalDate.of(2023, 10, 1), LocalDate.of(2023, 10, 31), SessionStatus.OPEN, 0);
     public static final Session SESSION3 = new Session(3L, 1L, "free", new FreeSessionType(), LocalDate.of(2023, 10, 1), LocalDate.of(2023, 10, 31), SessionStatus.CLOSED, 0);
