+++ src/test/java/nextstep/courses/domain/session/inner/ImageFormatTest.java
@@ -1,8 +1,7 @@
-package nextstep.courses.domain.session;
+package nextstep.courses.domain.session.inner;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
