+++ src/test/java/nextstep/courses/domain/session/inner/PaidSessionTypeTest.java
@@ -1,10 +1,11 @@
-package nextstep.courses.domain.session;
+package nextstep.courses.domain.session.inner;
 
 import static nextstep.courses.domain.session.SessionTest.SESSION2;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDate;
 
+import nextstep.courses.domain.session.Session;
 import nextstep.payments.domain.Payment;
 import org.junit.jupiter.api.Test;
